requirements.txt
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
marshmallow==3.14.0
pkg_resources==0.0.0
pyTelegramBotAPI==4.1.1
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tornado==6.1
tzdata==2021.2.post0
tzlocal==4.0
urllib3==1.26.7
